feat: parse add and remove actions

+++ src/actions/schemas.rs
@@ -77,9 +77,9 @@ fn add_fields() -> Vec<Field> {
             DataType::Struct(Fields::from(vec![
                 Field::new("storageType", DataType::Utf8, false),
                 Field::new("pathOrInlineDv", DataType::Utf8, false),
-                Field::new("offset", DataType::UInt64, true),
-                Field::new("sizeInBytes", DataType::UInt64, false),
-                Field::new("cardinality", DataType::UInt64, false),
+                Field::new("offset", DataType::Int64, true),
+                Field::new("sizeInBytes", DataType::Int32, false),
+                Field::new("cardinality", DataType::Int64, false),
             ])),
             true,
         ),
